requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT>=2.8.0
python-multipart>=0.0.9